+++ src/form/pari.cpp
@@ -180,8 +180,18 @@ bool PariFormula::eval(int64_t offset, int64_t numTerms, int timeoutSeconds,
   const std::string gpResult("pari-result-" + tmpFileId + ".txt");
   const int64_t maxparisize = 1024;  // in MB
   std::vector<std::string> args = {
-      "gp", "-s", std::to_string(maxparisize) + "M", "-q", gpPath};
+      "gp",
+      "-s",
+      std::to_string(maxparisize) + "M",
+      "--default",
+      "parisizemax=" + std::to_string(maxparisize) + "M",
+      "--default",
+      "recover=0",
+      "-q",
+      gpPath};
   std::string evalCode = printEvalCode(offset, numTerms);
+  // Provide "quit\n" on stdin to ensure PARI exits break loop cleanly on errors
   return SequenceUtil::evalFormulaWithExternalTool(
-      evalCode, getName(), gpPath, gpResult, args, timeoutSeconds, result);
+      evalCode, getName(), gpPath, gpResult, args, timeoutSeconds, result, "",
+      "quit\n");
 }